Move reactor db models into individual files
