+++ st2api/tests/unit/controllers/v1/test_triggerinstances.py
@@ -19,7 +19,7 @@ import six
 
 from st2common.transport.publishers import PoolPublisher
 from st2common.persistence.reactor import TriggerInstance
-from st2common.models.db.reactor import TriggerInstanceDB
+from st2common.models.db.trigger import TriggerInstanceDB
 from tests import FunctionalTest
 
 http_client = six.moves.http_client
