+++ st2common/st2common/models/api/rule.py
@@ -20,7 +20,7 @@ import six
 from st2common.models.api.base import BaseAPI
 from st2common.models.api.trigger import TriggerAPI
 from st2common.models.api.tag import TagsHelper
-from st2common.models.db.reactor import RuleDB, ActionExecutionSpecDB
+from st2common.models.db.rule import RuleDB, ActionExecutionSpecDB
 from st2common.models.system.common import ResourceReference
 from st2common.persistence.reactor import Trigger
 import st2common.services.triggers as TriggerService
