+++ st2common/tests/unit/test_reference.py
@@ -18,7 +18,7 @@ import mock
 import mongoengine
 
 from st2common.exceptions import db
-from st2common.models.db.reactor import TriggerDB
+from st2common.models.db.trigger import TriggerDB
 from st2common.persistence.reactor import Trigger
 from st2common.transport.publishers import PoolPublisher
 from st2common.util import reference
