+++ st2common/tests/unit/test_trigger_services.py
@@ -16,7 +16,7 @@
 from st2common.exceptions.triggers import TriggerDoesNotExistException
 from st2common.models.api.rule import RuleAPI
 from st2common.models.system.common import ResourceReference
-from st2common.models.db.reactor import TriggerDB
+from st2common.models.db.trigger import TriggerDB
 from st2common.persistence.reactor import (Trigger, TriggerType)
 import st2common.services.triggers as trigger_service
